requirements.txt
APScheduler==3.10.4
async-timeout==4.0.3
betterpath==0.2.2
click==8.1.7
greenlet==3.0.3
joblib==1.4.2
l==0.11.0
mysql-connector-python==9.0.0
numpy==2.1.0
pandas==2.2.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
SQLAlchemy==2.0.32
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
vcversioner==2.16.0.0
zope.interface==7.0.3
